requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.23.4
botocore==1.26.4
cloudinary==1.29.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.0.4
django-allauth==0.50.0
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.0
oauthlib==3.2.0
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
s3transfer==0.5.2
sqlparse==0.4.2
